+++ src/error.rs
@@ -7,7 +7,7 @@ pub(crate) enum Error {
     Toml(toml::de::Error),
     Utf8(std::str::Utf8Error),
     Utf8OsString,
-    MissingTarget,
+    MissingMuslTarget,
     MissingLicense,
 }
 
@@ -18,7 +18,7 @@ impl Display for Error {
             Error::Toml(e) => write!(f, "{}", e),
             Error::Utf8(e) => write!(f, "{}", e),
             Error::Utf8OsString => write!(f, "The `OsString` was not UTF-8!"),
-            Error::MissingTarget => write!(
+            Error::MissingMuslTarget => write!(
                 f,
                 "Missing target! Try: rustup target add x86_64-unknown-linux-musl"
             ),